Completely replaced old control.lua
